requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
reportlab>=4.0.0
pandas>=2.0.0
plotly>=5.0.0
requests>=2.31.0
pydantic-ai>=0.0.1 